Breaking circular dependence of shared_ptrs between arcs and ports.  Ports now store weak ptrs to arcs.  The arcs maintain ownership of nodes.  Arcs remove themselves from ports when their destructors are called.  This is accomplished by Arcs now containing a weak_ptr to themselves which is set in the factory.  Arcs are now owned by the design.
New ownership semantics mean that nodes will be kept so long as they are the child of a node*, are pointed to by an arc, or are contained in the design.  Note that parent nodes may be deleted causing a cascade of deletion.

Arcs will be kept so long as they are contained by the design.

These new semantics allow the graph objects to be freed when the associated design is deleted.  If another design references the same nodes and arcs, they will be kept.