Primarily fixing issues with topological scheduling.
Fixed a corner case in topological scheduling where the output is only connected to state elements.  In this case, it was not scheduled because the output arcs from state elements are removed for scheduling.

Fixed an issue with the output master not being scheduled since it was removed from the candidate set in the scheduling function.

Fixed an issue with pruning where subsystems were removed since they had no outdegree themselves.  Fixed an analogous bug in the topological scheduler which would attempt to schedule subsystems.

Added an exception for nodes with state in topological ordering validation when they are the src nodes.  This is because the

Also includes refactoring of function to get connected nodes to separate getting the connected input nodes and connected output node.

Changed default scheduling value to -1 (unscheduled)

Fixed an issue with indegree calculation when ignoring certain nodes.

Some various comment changes.