Fixed additional ClockDomain validation bug.  Fixed an issue where, durring ClockDomain specialization, the ClockDomain references in MasterNode ports were not updated.  This caused problems in the Design copy function.
Added a check to make sure the output directory exists before emitting the design.  This resolves an error in XercesC opening a GraphML file to dump the schedule graph when the output directory did not exist.  Added boost as a dependency to use cross platform filesystem implementation.

Also, restoring some try/catch blocks in main() functions.