In preparation for CircularBuffer in TappedDelay, changing the cEmit functions to return CExpr objects instead of simple strings. Added additional information to CExpr to indicate the type of expression, specifically scalar expression, scalar variable, array, or circular buffer. CExpr was also augmented with functions for emitting the correct indexing operation depending on whether the expression was an array or circular buffer.
Also inserted bugfix which applied to #56.  For nodes which acted as simple passthroughs (ex. ComplexToRealImag), it was possible for a
state variable to be passed through un-copied.  Since StateUpdate nodes are currently only dependent on direct consumers of a state node,
it was possible for the state to be updated before the value was used.  A temporary safeguard for this is to create temporary variables
for the outputs of these nodes.  This was already done for vector/matrix values but not for scalars.  Scalar temporary creation was added to:
- EnableInput.cpp
- StateUpdate.cpp
- DownsampleInput.cpp
- ComplexToRealImag.cpp
- Delay.cpp
- RealImagToComplex.cpp
- ReinterpretCast.cpp
- BlackBox.cpp – Comment Inserted as new logic not currently required but could be in the future